+++ Scouting/src/main/java/ca/team2706/scouting/mcmergemanager/gui/PreMatchReportFragment.java
@@ -12,7 +12,7 @@ import android.widget.TextView;
 
 import java.util.Map;
 
-import ca.team2706.scouting.mcmergemanager.miscBackend.StatsEngine;
+import ca.team2706.scouting.mcmergemanager.stronghold2016.dataObjects.StatsEngine;
 import ca.team2706.scouting.mcmergemanager.stronghold2016.dataObjects.MatchSchedule;
 import ca.team2706.scouting.mcmergemanager.R;
 import ca.team2706.scouting.mcmergemanager.stronghold2016.dataObjects.TeleopScoutingObject;
