+++ Scouting/src/main/java/ca/team2706/scouting/mcmergemanager/stronghold2016/gui/TeamStatsActivity.java
@@ -8,7 +8,7 @@ import android.widget.RelativeLayout;
 import android.widget.TextView;
 
 import ca.team2706.scouting.mcmergemanager.R;
-import ca.team2706.scouting.mcmergemanager.miscBackend.StatsEngine;
+import ca.team2706.scouting.mcmergemanager.stronghold2016.dataObjects.StatsEngine;
 import ca.team2706.scouting.mcmergemanager.stronghold2016.dataObjects.BallShot;
 
 import static ca.team2706.scouting.mcmergemanager.stronghold2016.dataObjects.TeleopScoutingObject.*;
